this.setState Automatically Calls render
A widow was in that city, and she often came to him, saying, 'Defend me from my adversary!'(Luke 18:3)
And there was a widow in that city; and she came unto him, saying, Avenge me of mine adversary.
THIS.STATE
this.setState Automatically Calls render
There's something odd about all of this.

Look again at Toggle.js.

When a user clicks on the <button></button>, the .changeColor() method is called. Take a look at .changeColor()'s definition.

.changeColor() calls this.setState(), which updates this.state.color. However, even if this.state.color is updated from green to yellow, that alone shouldn't be enough to make the screen's color change!

The screen's color doesn't change until Toggle renders.

Inside of the render function, it's this line:

<div style={{background:this.state.color}}>
that changes a virtual DOM object's color to the new this.state.color, eventually causing a change in the screen.

If you call .changeColor(), shouldn't you then also have to call .render() again? .changeColor() only makes it so that, the next time that you render, the color will be different. Why can you see the new background right away, if you haven't re-rendered the component?

Here's why: Any time that you call this.setState(), this.setState() AUTOMATICALLY calls .render() as soon as the state has changed.

Think of this.setState() as actually being two things: this.setState(), immediately followed by .render().

That is why you can't call this.setState() from inside of the .render() method! this.setState() automatically calls .render(). If .render() calls this.setState(), then an infinite loop is created.